requirements.txt
torch>=2.1.0
transformers>=4.42.0
accelerate
bitsandbytes
peft
loralib
einops

datasets
tqdm
scikit-learn

fastapi
uvicorn[standard]

pydantic

pyyaml